+++ mne_bids/path.py
@@ -953,8 +953,11 @@ class BIDSPath(object):
             )
             er_bids_path = _find_matched_empty_room(self)
 
-        if er_bids_path is not None:
-            assert er_bids_path.fpath.exists()
+        if er_bids_path is not None and not er_bids_path.fpath.exists():
+            raise FileNotFoundError(
+                f'Empty-room BIDS path resolved but not found:\n'
+                f'{er_bids_path}\n'
+                'Check your BIDS dataset for completeness.')
 
         return er_bids_path
 
@@ -1510,9 +1513,10 @@ def _find_matching_sidecar(bids_path, suffix=None,
                f'associated with {bids_path.basename}.')
     elif len(best_candidates) > 1:
         # More than one candidates were tied for best match
-        msg = (f'Expected to find a single {suffix} file '
+        msg = (f'Expected to find a single {search_suffix} file '
                f'associated with {bids_path.basename}, '
-               f'but found {len(candidate_list)}: "{candidate_list}".')
+               f'but found {len(candidate_list)}:\n\n' +
+               "\n".join(candidate_list))
     msg += f'\n\nThe search_str was "{search_str_complete}"'
     if on_error == 'raise':
         raise RuntimeError(msg)